Refactor, moving interfaces and more
