+++ OsuPlayer.IO/DbReader/DataModels/DbMapEntryBase.cs
@@ -1,5 +1,6 @@
 ﻿using System.Text;
 using Nein.Extensions;
+using OsuPlayer.IO.DbReader.Interfaces;
 
 namespace OsuPlayer.IO.DbReader.DataModels;
