+++ OsuPlayer.IO/DbReader/DataModels/RealmMapEntryBase.cs
@@ -1,5 +1,6 @@
 ﻿using System.Text;
 using Nein.Extensions;
+using OsuPlayer.IO.DbReader.Interfaces;
 using OsuPlayer.IO.Storage.LazerModels.Beatmaps;
 using OsuPlayer.IO.Storage.LazerModels.Files;
 using Realms;
