+++ OsuPlayer.IO/Storage/Blacklist/Blacklist.cs
@@ -1,6 +1,7 @@
 ﻿using Newtonsoft.Json;
 using OsuPlayer.Data.OsuPlayer.StorageModels;
 using OsuPlayer.IO.DbReader.DataModels;
+using OsuPlayer.IO.DbReader.Interfaces;
 
 namespace OsuPlayer.IO.Storage.Blacklist;
