+++ OsuPlayer/Modules/Audio/Interfaces/ICommonFeatures.cs
@@ -1,7 +1,7 @@
 ﻿namespace OsuPlayer.Modules.Audio.Interfaces;
 
 /// <summary>
-/// This interface is used to provide a common feature set used by by the audio component.
+/// This interface is used to provide a common feature set used by the audio component.
 /// </summary>
 public interface ICommonFeatures : IAudioControls, ISettableAudioDevices, IHasEffects
 {
