+++ OsuPlayer/Modules/Services/ISortProvider.cs
@@ -1,5 +1,6 @@
 ﻿using DynamicData;
 using OsuPlayer.Data.OsuPlayer.Enums;
+using OsuPlayer.IO.DbReader.Interfaces;
 
 namespace OsuPlayer.Modules.Services;
