+++ OsuPlayer/Views/ExportSongsViewModel.cs
@@ -2,6 +2,7 @@
 using Nein.Base;
 using Nein.Extensions;
 using OsuPlayer.Data.OsuPlayer.StorageModels;
+using OsuPlayer.IO.DbReader.Interfaces;
 using OsuPlayer.IO.Importer;
 using OsuPlayer.IO.Storage.Playlists;
 using ReactiveUI;
