+++ OsuPlayer/Views/HomeView.axaml.cs
@@ -3,6 +3,7 @@ using Avalonia.Interactivity;
 using Avalonia.Markup.Xaml;
 using Avalonia.VisualTree;
 using Nein.Base;
+using OsuPlayer.IO.DbReader.Interfaces;
 using OsuPlayer.IO.Storage.Blacklist;
 using OsuPlayer.UI_Extensions;
 using OsuPlayer.Windows;
