+++ OsuPlayer/Views/PlaylistEditorView.axaml
@@ -7,6 +7,7 @@
              xmlns:extensions="clr-namespace:OsuPlayer.Extensions.ValueConverters;assembly=OsuPlayer.Extensions"
              xmlns:views="clr-namespace:OsuPlayer.Views"
              xmlns:dataModels="clr-namespace:OsuPlayer.IO.DbReader.DataModels;assembly=OsuPlayer.IO"
+             xmlns:interfaces="clr-namespace:OsuPlayer.IO.DbReader.Interfaces;assembly=OsuPlayer.IO"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:Class="OsuPlayer.Views.PlaylistEditorView"
              FontWeight="{DynamicResource DefaultFontWeight}">
@@ -147,7 +148,7 @@
                          Items="{Binding FilteredSongEntries}" SelectedItems="{Binding SelectedSongListItems}"
                          SelectionChanged="SongList_OnSelectionChanged" DoubleTapped="PlaySong">
                     <ListBox.ItemTemplate>
-                        <DataTemplate DataType="dataModels:IMapEntryBase">
+                        <DataTemplate DataType="interfaces:IMapEntryBase">
                             <Grid>
                                 <TextBlock Text="{Binding SongName}" FontWeight="SemiBold" />
                             </Grid>
@@ -166,7 +167,7 @@
                          Items="{Binding CurrentSelectedPlaylist.Songs, Converter={StaticResource PlaylistValueConverter}}"
                          SelectionChanged="Playlist_OnSelectionChanged" DoubleTapped="PlaySong">
                     <ListBox.ItemTemplate>
-                        <DataTemplate DataType="dataModels:IMapEntryBase">
+                        <DataTemplate DataType="interfaces:IMapEntryBase">
                             <Grid>
                                 <TextBlock Text="{Binding SongName}" FontWeight="SemiBold" />
                             </Grid>
