+++ OsuPlayer/Views/PlaylistEditorView.axaml.cs
@@ -6,6 +6,7 @@ using Avalonia.VisualTree;
 using Nein.Base;
 using Nein.Extensions;
 using OsuPlayer.Data.OsuPlayer.StorageModels;
+using OsuPlayer.IO.DbReader.Interfaces;
 using OsuPlayer.IO.Storage.Playlists;
 using OsuPlayer.UI_Extensions;
 using OsuPlayer.Windows;
