+++ OsuPlayer/Views/PlaylistView.axaml
@@ -6,6 +6,7 @@
              xmlns:dataModels="clr-namespace:OsuPlayer.IO.DbReader.DataModels;assembly=OsuPlayer.IO"
              xmlns:avalonia="clr-namespace:Material.Icons.Avalonia;assembly=Material.Icons.Avalonia"
              xmlns:storageModels="clr-namespace:OsuPlayer.Data.OsuPlayer.StorageModels;assembly=OsuPlayer.Data"
+             xmlns:interfaces="clr-namespace:OsuPlayer.IO.DbReader.Interfaces;assembly=OsuPlayer.IO"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:Class="OsuPlayer.Views.PlaylistView"
              FontWeight="{DynamicResource DefaultFontWeight}">
@@ -66,7 +67,7 @@
                     </ContextMenu>
                 </ListBox.ContextMenu>
                 <ListBox.ItemTemplate>
-                    <DataTemplate DataType="dataModels:IMapEntryBase">
+                    <DataTemplate DataType="interfaces:IMapEntryBase">
                         <Grid ColumnDefinitions="*, Auto" RowDefinitions="Auto, Auto" DoubleTapped="PlaySong"
                               Background="Transparent">
                             <TextBlock Grid.Row="0" Grid.Column="0" Text="{Binding TitleString}" FontSize="20"
