+++ OsuPlayer/Views/SearchView.axaml
@@ -4,6 +4,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:views="clr-namespace:OsuPlayer.Views"
              xmlns:dataModels="clr-namespace:OsuPlayer.IO.DbReader.DataModels;assembly=OsuPlayer.IO"
+             xmlns:interfaces="clr-namespace:OsuPlayer.IO.DbReader.Interfaces;assembly=OsuPlayer.IO"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:Class="OsuPlayer.Views.SearchView"
              FontWeight="{DynamicResource DefaultFontWeight}">
@@ -34,7 +35,7 @@
                 </ContextMenu>
             </ListBox.ContextMenu>
             <ListBox.DataTemplates>
-                <DataTemplate DataType="dataModels:IMapEntryBase">
+                <DataTemplate DataType="interfaces:IMapEntryBase">
                     <Grid ColumnDefinitions="*, Auto" RowDefinitions="Auto, Auto">
                         <TextBlock Grid.Row="0" Grid.Column="0" Text="{Binding TitleString}" FontSize="20"
                                    FontWeight="SemiBold" />
